requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
base58==2.1.1
bitarray==3.8.0
bitcoin==1.1.42
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
ckzg==2.1.5
coincurve==21.0.0
constantly==23.10.4
cryptography==46.0.3
cytoolz==1.1.0
daphne==4.2.1
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
ecdsa==0.19.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonalias==0.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
packaging==25.0
parsimonious==0.10.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-bitcoinlib==0.12.2
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
rpds-py==0.30.0
service-identity==24.2.0
six==1.17.0
solders==0.27.1
sqlparse==0.5.5
toolz==1.1.0
tronpy==0.6.1
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.2
zope.interface==8.1.1
